[kingdom-backend-api] restore routes and migrate health endpoints
