+++ docs/patch_logs/backend_api_20250801164151.md
@@ -1,0 +1,8 @@
+# Backend API Patch Log - 2025-08-01
+
+## Summary
+- Re-enabled routers and cleaned up legacy routes.
+- Removed deprecated `/health` and `/metrics` handlers from `main.py`.
+- Moved database maintenance endpoints into `admin` router.
+- Added readiness and liveness probes to the `system` router.
+- Marked legacy `/api/*` and `/auth/*` stubs as excluded from OpenAPI.
